+++ crates/icondiffbot2/src/table_builder.rs
@@ -1,73 +1,6 @@
-use std::{
-    collections::{hash_map::DefaultHasher, HashMap},
-    hash::{Hash, Hasher},
-};
-
-use anyhow::{Context, Result};
-use diffbot_lib::{
-    github::{
-        github_api::download_url,
-        github_types::{CheckOutputs, ModifiedFileStatus, Output},
-    },
-    job::types::Job,
-};
-use dmm_tools::dmi::IconFile;
-
-pub fn status_to_sha(job: &Job, status: ModifiedFileStatus) -> (Option<&str>, Option<&str>) {
-    match status {
-        ModifiedFileStatus::Added => (None, Some(&job.head.sha)),
-        ModifiedFileStatus::Removed => (Some(&job.base.sha), None),
-        ModifiedFileStatus::Modified => (Some(&job.base.sha), Some(&job.head.sha)),
-        ModifiedFileStatus::Renamed => (None, None),
-        ModifiedFileStatus::Copied => (None, None),
-        ModifiedFileStatus::Changed => (None, None), // TODO: look up what this is
-        ModifiedFileStatus::Unchanged => (None, None),
-    }
-}
-
-pub struct IconFileWithName {
-    pub full_name: String,
-    pub sha: String,
-    pub hash: u64,
-    pub icon: IconFile,
-}
-
-pub async fn get_if_exists(
-    job: &Job,
-    filename: &str,
-    sha: Option<&str>,
-) -> Result<Option<IconFileWithName>> {
-    if let Some(sha) = sha {
-        let raw = download_url(&job.installation, &job.base.repo, filename, sha)
-            .await
-            .with_context(|| format!("Failed to download file {:?}", filename))?;
-
-        let mut hasher = DefaultHasher::new();
-        raw.hash(&mut hasher);
-        let hash = hasher.finish();
-
-        Ok(Some(IconFileWithName {
-            full_name: filename.to_string(),
-            sha: sha.to_string(),
-            hash,
-            icon: IconFile::from_raw(raw)
-                .with_context(|| format!("IconFile::from_raw failed for {:?}", filename))?,
-        }))
-    } else {
-        Ok(None)
-    }
-}
-
-pub async fn sha_to_iconfile(
-    job: &Job,
-    filename: &str,
-    sha: (Option<&str>, Option<&str>),
-) -> Result<(Option<IconFileWithName>, Option<IconFileWithName>)> {
-    Ok((
-        get_if_exists(job, filename, sha.0).await?,
-        get_if_exists(job, filename, sha.1).await?,
-    ))
-}
+use anyhow::Result;
+use diffbot_lib::github::github_types::{CheckOutputs, Output};
+use std::collections::HashMap;
 
 #[derive(Default, Debug)]
 pub struct OutputTableBuilder<'a> {
